Move Groupidentifier to right package, add apply for convenience
